requirements.txt
flask
flask-sqlalchemy
passlib
pyjwt

